+++ foxy_gh_farmer/foxy_gh_farmer_main.py
@@ -6,6 +6,7 @@ from chia.cmds.passphrase import passphrase_cmd
 
 from foxy_gh_farmer.cmds.farm_summary import summary_cmd
 from foxy_gh_farmer.cmds.authenticate import authenticate_cmd
+from foxy_gh_farmer.cmds.init import init_cmd
 from foxy_gh_farmer.cmds.join_pool import join_pool_cmd
 from foxy_gh_farmer.error_reporting import close_sentry
 from foxy_gh_farmer.util.root_path import get_root_path
@@ -62,6 +63,7 @@ cli.add_command(join_pool_cmd)
 cli.add_command(authenticate_cmd)
 cli.add_command(keys_cmd)
 cli.add_command(passphrase_cmd)
+cli.add_command(init_cmd)
 
 
 def main() -> None: